feat : create dbclient by wire
