+++ internal/config/config.go
@@ -1,10 +1,5 @@
 package config
 
-import (
-   "gorm.io/gorm"
-	"gorm.io/driver/mysql"
-)
-
 var RuntimeConf = RuntimeConfig{}
 
 type RuntimeConfig struct {
@@ -31,15 +26,3 @@ type Server struct {
    Port int `yaml:"port"`
 }
 
-func ConnectDb() *gorm.DB {
-	DatasourceUrl := RuntimeConf.Datasource.Db.Url
-	UserName := RuntimeConf.Datasource.Db.UserName
-	Password := RuntimeConf.Datasource.Db.Password
-	dsn := UserName + ":" + Password + "@" + DatasourceUrl
-	
-	db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
-	if err != nil {
-	  panic("Db 연결에 실패하였습니다.")
-	}
-	return db
-}
